+++ Week3/assignment/src/utils/getRandomNumber.js
@@ -1,0 +1,10 @@
+export function getRandomNumber() {
+  const randomNumbers = [];
+  while (randomNumbers.length < 3) {
+    const random = Math.floor(Math.random() * 10);
+    if (!randomNumbers.includes(random)) {
+      randomNumbers.push(random);
+    }
+  }
+  return randomNumbers;
+}
